hosts.py: Added ALLOWED_HOSTS lazy iterable object.
By default, Django 1.5 requires ALLOWED_HOSTS to be defined when DEBUG is
False. This lazy iterable object allows us to query our aliases for hosts.

We can also specify wild cards by setting `MULTISITE_EXTRA_HOSTS` if we so
desire. The hosts specified in this variable will be verified first so we
don't have to hit the database if we don't have to.

To use, put the following in your Django settings file:
  from multisite.hosts import ALLOWED_HOSTS as AH
  ALLOWED_HOSTS = AH

If you want to specify extra allowed hosts, add this line to your settings:
  MULTISITE_EXTRA_HOSTS = ['*.mydomain.com']