Fix "error reading TLD cache file" error
The cache_file argument to TLDExtract was an empty tempfile,
so pickle complained about an EOF error when loading it.
That was then logged as an "error reading TLD cache file"
error by tldextract.

Contrary to an empty cache_file, tldextract deals with a non-existent
cache_file just fine, so we don't really need that intermediate
step of creating a tempfile anyway. Use the TLDExtract `update` method
to directly update multisite's public suffix list