Repair under Django 1.6 / python2
This is an improved @jordiecometrica's b57a470ffa796bbf75cf83fd870445763f2ec1a0
which special-cases the 1.6 branch.

Note: our setup.py says we're only supporting back to Django 1.6.
An equally good way to handle this would be to bump that to 1.7,
and given that not even 1.7 is getting security patches anymore,
maybe we should just go straight to 1.8.
But for now, be conservative.